+++ src/test/index.js
@@ -1,0 +1,3 @@
+import renderWithTheme from './renderWithTheme'
+
+export { renderWithTheme }